+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 2.16
 ----
 
-* Not yet released.
+* Released on 20th December 2023.
 * Fixed detection UTF-16 Strings files.
 * Fixed detection of files with some special chars.
 * Added support for Mobile Kotlin resources.