+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+2.17
+----
+
+* Not yet released.
+
 2.16
 ----